requirements.txt
Flask==2.0.1
Werkzeug==2.0.1 
pandas==1.1.5 
Pillow==8.3.1
gunicorn==20.1.0
